refactor(config): move config types to shared types file
Consolidate all configuration type definitions into a single types file for better maintainability and to reduce duplication. This includes moving PaymentConfig, ThemeConfig, and FeaturesConfig interfaces to src/types/index.d.ts.

Also update README.md to reflect current setup instructions and remove unused admin setup step.
