+++ src/config/features.config.ts
@@ -1,75 +1,5 @@
-export interface FeaturesConfig {
-  auth: {
-    enabled: boolean;
-    providers: {
-      email: boolean;
-      github: boolean;
-      google: boolean;
-    };
-    session: {
-      maxAge: number;
-    };
-    passwordReset: boolean;
-    emailVerification: boolean;
-  };
-  payment: {
-    enabled: boolean;
-    provider: 'stripe';
-    currency: string;
-    trial: {
-      enabled: boolean;
-      days: number;
-    };
-  };
-  fileManager: {
-    enabled: boolean;
-    storage: 'r2' | 's3';
-    thumbnails: boolean;
-    imageProcessing: boolean;
-    maxFileSize: number;
-    allowedTypes: string[];
-  };
-  blog: {
-    enabled: boolean;
-    commentsEnabled: boolean;
-    tagsEnabled: boolean;
-    authorsEnabled: boolean;
-    searchEnabled: boolean;
-  };
-  docs: {
-    enabled: boolean;
-    searchEnabled: boolean;
-    editOnGithub: boolean;
-    tableOfContents: boolean;
-    breadcrumbs: boolean;
-  };
-  analytics: {
-    enabled: boolean;
-    provider: 'vercel' | 'google' | 'plausible';
-    trackingId?: string;
-  };
-  notifications: {
-    enabled: boolean;
-    emailNotifications: boolean;
-    pushNotifications: boolean;
-    inAppNotifications: boolean;
-  };
-  dashboard: {
-    enabled: boolean;
-    widgets: {
-      analytics: boolean;
-      recentActivity: boolean;
-      quickActions: boolean;
-      notifications: boolean;
-    };
-  };
-  admin: {
-    enabled: boolean;
-    userManagement: boolean;
-    systemSettings: boolean;
-    analytics: boolean;
-  };
-}
+import type { FeaturesConfig } from "@/types";
+
 
 export const featuresConfig: FeaturesConfig = {
   // Authentication features
