+++ src/config/index.ts
@@ -4,10 +4,3 @@ export { featuresConfig } from './features.config';
 export { i18nConfig } from './i18n.config';
 export { themeConfig } from './theme.config';
 export { paymentConfig } from './payment.config';
-
-// Export types
-export type { AppConfig } from './app.config';
-export type { FeaturesConfig } from './features.config';
-export type { I18nConfig } from './i18n.config';
-export type { ThemeConfig } from './theme.config';
-export type { PaymentConfig } from './payment.config';
