+++ src/config/payment.config.ts
@@ -1,62 +1,4 @@
-import type { PaymentInterval } from '@/payment/types';
-
-export interface PaymentPlan {
-  id: string;
-  name: string;
-  description: string;
-  price: number;
-  interval: PaymentInterval;
-  stripePriceId?: string;
-  stripePriceIds?: {
-    monthly?: string;
-    yearly?: string;
-  };
-  // Add yearly price for plans that support both monthly and yearly billing
-  yearlyPrice?: number;
-  features: string[];
-  popular?: boolean;
-  metadata?: Record<string, string>;
-  limits?: {
-    storage?: number; // in GB
-    users?: number;
-    projects?: number;
-    apiCalls?: number;
-  };
-}
-
-export interface PaymentConfig {
-  provider: 'stripe';
-  currency: string;
-  stripe: {
-    publishableKey: string;
-    secretKey: string;
-    webhookSecret: string;
-    apiVersion: string;
-  };
-  plans: PaymentPlan[];
-  trial: {
-    enabled: boolean;
-    days: number;
-    plans: string[]; // plan IDs that support trial
-  };
-  invoice: {
-    footer: string;
-    logo?: string;
-    supportEmail: string;
-  };
-  billing: {
-    collectTaxId: boolean;
-    allowPromotionCodes: boolean;
-    automaticTax: boolean;
-  };
-  features: {
-    subscriptions: boolean;
-    oneTimePayments: boolean;
-    invoices: boolean;
-    customerPortal: boolean;
-    webhooks: boolean;
-  };
-}
+import type { PaymentConfig } from '@/types';
 
 export const paymentConfig: PaymentConfig = {
   // Payment provider
