+++ src/config/theme.config.ts
@@ -1,73 +1,4 @@
-export interface ThemeConfig {
-  defaultTheme: 'light' | 'dark' | 'system';
-  themes: readonly string[];
-  colors: {
-    primary: Record<string, string>;
-    secondary: Record<string, string>;
-    accent: Record<string, string>;
-    neutral: Record<string, string>;
-    success: Record<string, string>;
-    warning: Record<string, string>;
-    error: Record<string, string>;
-    info: Record<string, string>;
-  };
-  fonts: {
-    sans: string[];
-    mono: string[];
-    serif: string[];
-  };
-  borderRadius: {
-    none: string;
-    sm: string;
-    md: string;
-    lg: string;
-    xl: string;
-    full: string;
-  };
-  spacing: {
-    xs: string;
-    sm: string;
-    md: string;
-    lg: string;
-    xl: string;
-    '2xl': string;
-    '3xl': string;
-  };
-  animations: {
-    duration: {
-      fast: string;
-      normal: string;
-      slow: string;
-    };
-    easing: {
-      ease: string;
-      easeIn: string;
-      easeOut: string;
-      easeInOut: string;
-    };
-  };
-  breakpoints: {
-    sm: string;
-    md: string;
-    lg: string;
-    xl: string;
-    '2xl': string;
-  };
-  shadows: {
-    sm: string;
-    md: string;
-    lg: string;
-    xl: string;
-    '2xl': string;
-  };
-  zIndex: {
-    dropdown: number;
-    modal: number;
-    popover: number;
-    tooltip: number;
-    toast: number;
-  };
-}
+import type { ThemeConfig } from "@/types";
 
 export const themeConfig: ThemeConfig = {
   // Default theme
