+++ src/types/index.d.ts
@@ -1,3 +1,5 @@
+import type { PaymentInterval } from '@/payment/types';
+
 export interface AppConfig {
   app: {
     name: string;
@@ -72,6 +74,209 @@ export interface I18nConfig {
     percent: Intl.NumberFormatOptions;
   }>;
 }
+
+export interface FeaturesConfig {
+  auth: {
+    enabled: boolean;
+    providers: {
+      email: boolean;
+      github: boolean;
+      google: boolean;
+    };
+    session: {
+      maxAge: number;
+    };
+    passwordReset: boolean;
+    emailVerification: boolean;
+  };
+  payment: {
+    enabled: boolean;
+    provider: 'stripe';
+    currency: string;
+    trial: {
+      enabled: boolean;
+      days: number;
+    };
+  };
+  fileManager: {
+    enabled: boolean;
+    storage: 'r2' | 's3';
+    thumbnails: boolean;
+    imageProcessing: boolean;
+    maxFileSize: number;
+    allowedTypes: string[];
+  };
+  blog: {
+    enabled: boolean;
+    commentsEnabled: boolean;
+    tagsEnabled: boolean;
+    authorsEnabled: boolean;
+    searchEnabled: boolean;
+  };
+  docs: {
+    enabled: boolean;
+    searchEnabled: boolean;
+    editOnGithub: boolean;
+    tableOfContents: boolean;
+    breadcrumbs: boolean;
+  };
+  analytics: {
+    enabled: boolean;
+    provider: 'vercel' | 'google' | 'plausible';
+    trackingId?: string;
+  };
+  notifications: {
+    enabled: boolean;
+    emailNotifications: boolean;
+    pushNotifications: boolean;
+    inAppNotifications: boolean;
+  };
+  dashboard: {
+    enabled: boolean;
+    widgets: {
+      analytics: boolean;
+      recentActivity: boolean;
+      quickActions: boolean;
+      notifications: boolean;
+    };
+  };
+  admin: {
+    enabled: boolean;
+    userManagement: boolean;
+    systemSettings: boolean;
+    analytics: boolean;
+  };
+}
+
+export interface ThemeConfig {
+  defaultTheme: 'light' | 'dark' | 'system';
+  themes: readonly string[];
+  colors: {
+    primary: Record<string, string>;
+    secondary: Record<string, string>;
+    accent: Record<string, string>;
+    neutral: Record<string, string>;
+    success: Record<string, string>;
+    warning: Record<string, string>;
+    error: Record<string, string>;
+    info: Record<string, string>;
+  };
+  fonts: {
+    sans: string[];
+    mono: string[];
+    serif: string[];
+  };
+  borderRadius: {
+    none: string;
+    sm: string;
+    md: string;
+    lg: string;
+    xl: string;
+    full: string;
+  };
+  spacing: {
+    xs: string;
+    sm: string;
+    md: string;
+    lg: string;
+    xl: string;
+    '2xl': string;
+    '3xl': string;
+  };
+  animations: {
+    duration: {
+      fast: string;
+      normal: string;
+      slow: string;
+    };
+    easing: {
+      ease: string;
+      easeIn: string;
+      easeOut: string;
+      easeInOut: string;
+    };
+  };
+  breakpoints: {
+    sm: string;
+    md: string;
+    lg: string;
+    xl: string;
+    '2xl': string;
+  };
+  shadows: {
+    sm: string;
+    md: string;
+    lg: string;
+    xl: string;
+    '2xl': string;
+  };
+  zIndex: {
+    dropdown: number;
+    modal: number;
+    popover: number;
+    tooltip: number;
+    toast: number;
+  };
+}
+
+export interface PaymentPlan {
+  id: string;
+  name: string;
+  description: string;
+  price: number;
+  interval: PaymentInterval;
+  stripePriceId?: string;
+  stripePriceIds?: {
+    monthly?: string;
+    yearly?: string;
+  };
+  // Add yearly price for plans that support both monthly and yearly billing
+  yearlyPrice?: number;
+  features: string[];
+  popular?: boolean;
+  metadata?: Record<string, string>;
+  limits?: {
+    storage?: number; // in GB
+    users?: number;
+    projects?: number;
+    apiCalls?: number;
+  };
+}
+
+export interface PaymentConfig {
+  provider: 'stripe';
+  currency: string;
+  stripe: {
+    publishableKey: string;
+    secretKey: string;
+    webhookSecret: string;
+    apiVersion: string;
+  };
+  plans: PaymentPlan[];
+  trial: {
+    enabled: boolean;
+    days: number;
+    plans: string[]; // plan IDs that support trial
+  };
+  invoice: {
+    footer: string;
+    logo?: string;
+    supportEmail: string;
+  };
+  billing: {
+    collectTaxId: boolean;
+    allowPromotionCodes: boolean;
+    automaticTax: boolean;
+  };
+  features: {
+    subscriptions: boolean;
+    oneTimePayments: boolean;
+    invoices: boolean;
+    customerPortal: boolean;
+    webhooks: boolean;
+  };
+}
+
 
 export interface SidebarItem {
   title: string;
